Refactor BnfProxyController to use BnfService for searching books and extract related logic into a new service class
