+++ app/Services/BnfService.php
@@ -1,0 +1,87 @@
+<?php
+namespace App\Services;
+
+use Illuminate\Support\Facades\Http;
+use App\Models\Book;
+use App\Services\BookCacheService;
+
+class BnfService
+{
+    public function search(string $query, int $page = 1): array
+    {
+        $isIsbn = preg_match('/^(?:\d{9}[\dX]|\d{13})$/', $query);
+        $finalQuery = $isIsbn
+            ? 'isbn="' . $query . '"'
+            : 'bib.doctype all "a" AND title="' . $query . '" AND bib.language any "fre"';
+
+        $params = [
+            'version' => '1.2',
+            'operation' => 'searchRetrieve',
+            'query' => $finalQuery,
+            'startRecord' => 1 + ($page - 1) * 30,
+            'maximumRecords' => $isIsbn ? 1 : 30,
+            'recordSchema' => 'dc',
+        ];
+
+        $response = Http::withoutVerifying()->get('https://catalogue.bnf.fr/api/SRU', $params);
+        $xml = simplexml_load_string($response->body());
+
+        return $this->parseXml($xml);
+    }
+
+    private function parseXml($xml): array
+    {
+        $xml->registerXPathNamespace('srw', 'http://www.loc.gov/zing/srw/');
+        $records = $xml->xpath('//srw:record');
+        $books = [];
+
+        foreach ($records as $record) {
+            $dc = $record->children('http://www.loc.gov/zing/srw/')
+                         ->recordData
+                         ->children('http://www.openarchives.org/OAI/2.0/oai_dc/')
+                         ->dc;
+
+            if (!$dc) continue;
+
+            $isbn = $this->extractISBN($dc->children('http://purl.org/dc/elements/1.1/')->identifier);
+
+            if (empty($isbn) || $isbn === 'ISBN inconnu') continue;
+
+            $bookModel = Book::find($isbn);
+            if ($bookModel) {
+                $books[] = $bookModel;
+                continue;
+            }
+
+            $book = [
+                'title'     => $this->clean((string) $dc->children('http://purl.org/dc/elements/1.1/')->title),
+                'isbn'      => $isbn,
+                'author'    => $this->clean((string) $dc->children('http://purl.org/dc/elements/1.1/')->creator),
+                'year'      => (string) $dc->children('http://purl.org/dc/elements/1.1/')->date ?: 'Date inconnue',
+                'publisher' => $this->clean((string) $dc->children('http://purl.org/dc/elements/1.1/')->publisher),
+            ];
+
+            BookCacheService::store($book);
+            $books[] = $book;
+        }
+
+        return $books;
+    }
+
+    private function extractISBN($identifiers)
+    {
+        foreach ((array) $identifiers as $id) {
+            if (str_contains($id, 'ISBN')) {
+                return trim(str_replace('ISBN', '', $id));
+            }
+        }
+        return 'ISBN inconnu';
+    }
+
+    private function clean($text)
+    {
+        $text = preg_replace('/\(.*?\)|\s\/.*$|\s;.*$/', '', $text);
+        $text = preg_replace('/\. Auteur du texte/', '', $text);
+        return trim($text);
+    }
+}
